move drop code into its own file
